HHH-5917 - Split test cases and updated documentation
